[adbd] Support screenshot via gralloc framebuffer HAL
As there is a hardware abstraction layer for framebuffer access, the
adbd screenshot utility should make use of it instead of reading from
/dev/graphics/fb0. This removes the need for hardware vendors to
implement the kernel framebuffer driver.

Adbd, being a statically linked executable, can not directly depend on
libhardware for accessing the gralloc module. To circumvent this a new
tool called "fbread" is created to dump the current contents of the
framebuffer into stdout. For backwards compatibility the fbread
utility will fall back to reading from /dev/graphics/fb0 if the
gralloc implementation does not support the read()
operation. Additionally the gralloc read code is only compiled when
GRALLOC_FB_READ_SUPPORTED, for source level backwards compatibility
with older versions of the gralloc interface.

The old create_subprocess() routine in adb/services.c created a
pseudoterminal, the old routine is renamed to create_terminal() to
better describe its purpose and a simpler implementation of
create_subprocess() that redirects the child's streams into a socket
directly is added.

Change-Id: Ia810fbe6d86cffa5f531f52ef177ced2434a3222